Adjustable Refresh rate
Replaced for loop with graph function to implement graph drawing based on given interval. Basic layout for temp readings on screen with labels and axis placed correctly.  Rough count up of x (equals seconds) ok so far, since the offset to the given interval are in the regime of max 10 ms.
SD card logging needs to be implemented with defined starting point based on user interface.